refactor: restructure step2
