+++ src/components/common/styles.ts
@@ -1,0 +1,6 @@
+import { css } from '@emotion/css'
+
+export const styles = {
+  header: css({ textAlign: 'center'}),
+  footer: css({textAlign: 'center'})
+}
